requirements.txt
streamlit
pandas
PyMuPDF
python-docx
sentence-transformers
torch
Pillow
pytesseract
